back/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
idna==3.4
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
parso==0.8.3
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.21
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.15.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.13
